Display element has been created
Display element has been created  and imported to list.jsx where the main descison making is happening.
The next step is to destructure props and display the data in some way in Display. Also keys were added to the list while mapping in list.jsx to prevent the error. Other errors related to how css properties are named in inline styles in react fixed.