Exercises 2.18 and 2.19 are done
Exercises 2.18 and 2.19 are done. 2.19 wants us to do buttons so the listed options are clickable. I have finished it yesterday by accident, because it was making sense. I also added some styles just for fun, and since I didn't mess with css for a while a went a bit wild. Styles a vary row, but shall live it as is for now, there are still a lot to do before I can jump to section 3.